+++ src/main/kotlin/step3/IntRandom.kt
@@ -1,0 +1,12 @@
+package step3
+
+import java.util.Random
+
+interface IntRandom {
+    fun next(): Int
+    class Smart(private val random: Random, private val bound: Int = Int.MAX_VALUE) : IntRandom {
+        override fun next(): Int {
+            return random.nextInt(bound)
+        }
+    }
+}
