+++ src/main/kotlin/step3/Position.kt
@@ -1,0 +1,9 @@
+package step3
+
+data class Position(private val position: Int) {
+    init {
+        require(position >= 0)
+    }
+
+    fun increase() = copy(position = position + 1)
+}
